bring in line with internal version
This commit brings the public repository up to date with general changes
made to the version used on non-public projects.

Ascalon was primarily written over two years ago. There's a lot I can do
to overhaul it with the knowledge and experience I have gained since
then; and I may do so one day if I have the time, but for now it still
works well enough for my needs that I haven't needed to.

# Unity default UI
- added a button to scroll to the bottom of the entry list
- the input field is now automatically re-selected upon deselection

# Internal
- the config loader now catches exceptions while running a config file,
  so that an issue caused by one command or variable will not prevent
  execution of further lines
- disabled remaining commands/entries in TestCommands
- ConVars now have an internal function to set the data manually,
  intended primarily for use during their DataChanged functions (e.g.
  clamping values)